+++ pages/_app.tsx
@@ -1,15 +1,14 @@
-import "@/styles/globals.css";
-import type { AppProps } from "next/app";
-import { ThemeProvider } from "next-themes";
 import Layout from "@/sharedUI/landingUI/Layout";
-
+import { ThemeProvider } from "next-themes";
+import type { AppProps } from "next/app";
 import { IBM_Plex_Sans_Arabic } from "next/font/google";
 
+import "@/styles/globals.css";
 const IBMfont = IBM_Plex_Sans_Arabic({
   weight: ["100", "200", "300", "400", "500", "600", "700"],
   subsets: ["arabic"],
   display: "auto",
-  variable: "--font-ibm-plex-sans-arabic",
+  variable: "--font-ibm-plex-sans-arabic"
 });
 
 export default function App({ Component, pageProps }: AppProps) {